requirements.txt
alabaster==0.7.12
altgraph==0.16.1
atomicwrites==1.3.0
attrs==19.3.0
Babel==2.7.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
colorama==0.4.3
cryptography==2.8
Cython==0.29.14
docutils==0.15.2
future==0.18.2
idna==2.8
imagesize==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
more-itertools==8.0.2
packaging==19.2
pefile==2019.4.18
pluggy==0.13.1
py==1.8.0
pycparser==2.19
pycryptodome==3.9.4
Pygments==2.5.2
pyparsing==2.4.5
pytest==5.3.2
pytz==2019.3
pywin32-ctypes==0.2.0
regex==2020.7.14
requests==2.22.0
selenium==3.141.0
simplejson==3.17.0
six==1.13.0
snowballstemmer==2.0.0
Sphinx==2.3.0
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
Unidecode==1.1.1
urllib3==1.25.7
wcwidth==0.1.7